refactor: DetailQuizControllerTest 코틀린 변환
